# docs/swagger.yaml
basePath: /v1
definitions:
...
  handler.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  models.User:
    properties:
      created_at:
        type: string
      current_traffic:
        type: integer
      email:
        type: string
      id:
        type: integer
      max_traffic:
        type: integer
      subscription_token:
        type: string
      updated_at:
        type: string
      username:
        type: string
      uuid:
        type: string
    type: object
...
  users.destroyResponse:
    properties:
      user:
        type: string
    type: object
...
  users.showResponse:
    properties:
      user:
        $ref: '#/definitions/models.User'
        type: object
    type: object
...
  users.updateRequest:
    properties:
      email:
        type: string
      subscription_token:
        type: string
      uuid:
        type: string
    type: object
  users.updateResponse:
    properties:
      user:
        $ref: '#/definitions/models.User'
        type: object
    type: object
...
host: localhost
info:
  contact: {}
  description: A Swagger UI For RayDash API
  license:
    name: GPLv3
    url: https://www.gnu.org/licenses/gpl-3.0.html
  title: RayDash API
  version: 1.0.0
paths:
...
  /users/{username}:
    delete:
      consumes:
      - application/json
      description: Delete a user according to username
      operationId: users.Destroy
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      - description: Node Token
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.destroyResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete a user
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: Return user according to username in url
      operationId: users.Show
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      - description: Node Token
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.showResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Show the required user
      tags:
      - Users
    patch:
      consumes:
      - application/json
      description: Update the user provided
      operationId: users.Update
      parameters:
      - description: User Object
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/users.updateRequest'
      - description: Username
        in: path
        name: username
        required: true
        type: string
      - description: Node Token
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.updateResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update user
      tags:
      - Users
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"